requirements.txt
requests==2.27.1
pymongo==4.8.0
python-dotenv
Flask==2.1.3
Werkzeug==2.2.3
Flask-PyMongo==2.3.0
